# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/roles": {
      "post": {
        "operationId": "RolesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      },
      "get": {
        "operationId": "RolesController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/roles/{id}": {
      "get": {
        "operationId": "RolesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      },
      "delete": {
        "operationId": "RolesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/roles/{id}/name": {
      "patch": {
        "operationId": "RolesController_rename_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/roles/{id}/permissions": {
      "patch": {
        "operationId": "RolesController_updatePermissions_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRolePermissionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "roles"
        ]
      }
    },
...
    "/api/v1/athletes": {
      "post": {
        "operationId": "AthletesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAthleteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "athletes"
        ]
      },
      "get": {
        "operationId": "AthletesController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "athletes"
        ]
      }
    },
    "/api/v1/athletes/public": {
      "get": {
        "operationId": "AthletesController_findAllPublic_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-indexed page number.",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (max 200).",
            "schema": {
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "athletes"
        ]
      }
    },
    "/api/v1/athletes/{id}": {
      "get": {
        "operationId": "AthletesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "athletes"
        ]
      },
      "delete": {
        "operationId": "AthletesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "athletes"
        ]
      }
    },
...
    "/api/v1/albums/public/{slug}": {
      "get": {
        "operationId": "AlbumsController_getPublicBySlug_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The album detail page composite (album + visible photos + related albums). Literally `null` in the response body — still HTTP 200, not 404 — when no Published album matches `slug`; the caller is responsible for treating a null body as not-found (matches the `AthleteProfilesService` convention this endpoint was built from).",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/AlbumDetailPageResponseDto"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "albums"
        ]
      }
    },
...
    "/api/v1/pages": {
      "post": {
        "operationId": "PagesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "pages"
        ]
      },
      "get": {
        "operationId": "PagesController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "pages"
        ]
      }
    },
    "/api/v1/pages/public/{slug}": {
      "get": {
        "operationId": "PagesController_findPublishedBySlug_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/PagePublicResponseDto"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "pages"
        ]
      }
    },
    "/api/v1/pages/{id}": {
      "get": {
        "operationId": "PagesController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "pages"
        ]
      },
      "delete": {
        "operationId": "PagesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "pages"
        ]
      }
    },
    "/api/v1/page-sections": {
      "post": {
        "operationId": "PageSectionsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePageSectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "page-sections"
        ]
      },
      "get": {
        "operationId": "PageSectionsController_findAll_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "page-sections"
        ]
      }
    },
    "/api/v1/page-sections/public/by-page/{pageId}": {
      "get": {
        "operationId": "PageSectionsController_findPublicByPage_v1",
        "parameters": [
          {
            "name": "pageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PageSectionPublicResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "page-sections"
        ]
      }
    },
    "/api/v1/page-sections/{id}": {
      "get": {
        "operationId": "PageSectionsController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "page-sections"
        ]
      },
      "delete": {
        "operationId": "PageSectionsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "page-sections"
        ]
      }
    },
...
  },
  "info": {
    "title": "UAEAF Backend API",
    "description": "UAE Athletics Federation platform API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LocalizedTextDto": {
        "type": "object",
        "properties": {
          "en": {
            "type": "string",
            "description": "English text."
          },
          "ar": {
            "type": "string",
            "description": "Arabic text."
          }
        },
        "required": [
          "en",
          "ar"
        ]
      },
...
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "description": "Bilingual role name, e.g. { en: \"News Approver\", ar: \"معتمد الأخبار\" }.",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalizedTextDto"
              }
            ]
          },
          "description": {
            "description": "Bilingual explanation of what this role is for. Optional.",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalizedTextDto"
              }
            ]
          },
          "permissionIds": {
            "description": "Permission ids granted to this role.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "permissionIds"
        ]
      },
      "RenameRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "description": "New bilingual role name.",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalizedTextDto"
              }
            ]
          },
          "description": {
            "description": "Bilingual explanation of what the role is for. Omit to leave the stored description unchanged; send null to remove it.",
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalizedTextDto"
              }
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateRolePermissionsDto": {
        "type": "object",
        "properties": {
          "permissionIds": {
            "description": "Full replacement list of permission ids for this role.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "permissionIds"
        ]
      },
...
      "CreateAthleteDto": {
        "type": "object",
        "properties": {
          "name": {
            "description": "Bilingual athlete name.",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalizedTextDto"
              }
            ]
          },
          "dateOfBirth": {
            "type": "string"
          },
          "nationalityId": {
            "type": "string"
          },
          "disciplineIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "gender": {
            "type": "string",
            "enum": [
              "Male",
              "Female"
            ]
          },
          "residencyType": {
            "type": "string",
            "enum": [
              "Local",
              "Guest"
            ]
          },
          "federationName": {
            "description": "Home national federation display name — populated only when residencyType=Guest.",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalizedTextDto"
              }
            ]
          }
        },
        "required": [
          "name",
          "dateOfBirth",
          "nationalityId",
          "gender",
          "residencyType"
        ]
      },
...
      "AlbumPublicResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "slug": {
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalizedTextDto"
              }
            ]
          },
          "contentCategoryId": {
            "type": "string"
          },
          "championshipName": {
            "nullable": true,
            "description": "Denormalized championship/tournament name, captured at album-creation time — not a live join (the championships collection does not exist yet). [REVIEW WHEN THE CHAMPIONSHIPS MODULE IS BUILT].",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalizedTextDto"
              }
            ]
          },
          "coverImageId": {
            "type": "object",
            "nullable": true
          },
          "publishedAt": {
            "type": "object",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "assetCount": {
            "type": "number",
            "description": "Denormalized visible-photo count."
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "contentCategoryId",
          "tags",
          "assetCount"
        ]
      },
...
      "MediaAssetPublicResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "file": {
            "$ref": "#/components/schemas/MediaFilePublicResponseDto"
          },
          "caption": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "altText": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "displayOrder": {
            "type": "number"
          },
          "isFeatured": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "file",
          "caption",
          "altText",
          "displayOrder",
          "isFeatured"
        ]
      },
      "RelatedAlbumSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "slug": {
            "type": "string"
          },
          "coverImageId": {
            "type": "object",
            "nullable": true
          },
          "publishedAt": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "title",
          "slug"
        ]
      },
      "AlbumDetailPageResponseDto": {
        "type": "object",
        "properties": {
          "album": {
            "$ref": "#/components/schemas/AlbumPublicResponseDto"
          },
          "mediaAssets": {
            "description": "Visible photos, in display order.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaAssetPublicResponseDto"
            }
          },
          "relatedAlbums": {
            "description": "Other Published albums sharing an association target, excluding this one.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelatedAlbumSummaryDto"
            }
          }
        },
        "required": [
          "album",
          "mediaAssets",
          "relatedAlbums"
        ]
      },
...
      "PagePublicResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "seo": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PageSeoPublicResponseDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "slug",
          "title"
        ]
      },
      "PageSeoDto": {
        "type": "object",
        "properties": {
          "metaTitle": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "metaDescription": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "ogImageId": {
            "type": "string"
          }
        }
      },
      "CreatePageDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "Unique route slug."
          },
          "title": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "status": {
            "type": "string",
            "enum": [
              "Draft",
              "Published"
            ],
            "description": "Structural routing status, not publicationState."
          },
          "seo": {
            "$ref": "#/components/schemas/PageSeoDto"
          }
        },
        "required": [
          "slug",
          "title",
          "status"
        ]
      },
      "CreatePageSectionDto": {
        "type": "object",
        "properties": {
          "pageId": {
            "type": "string"
          },
          "sectionType": {
            "type": "string",
            "enum": [
              "HERO",
              "FEDERATION_STATS",
              "FEATURED_ATHLETES",
              "RESULTS_RANKINGS",
              "CLUBS",
              "UPCOMING_HIGHLIGHTS",
              "LATEST_NEWS",
              "PHOTO_GALLERY",
              "VIDEO_LIBRARY",
              "EXTERNAL_MEDIA",
              "SPONSORS",
              "PARTNERS",
              "NEWSLETTER_CTA"
            ]
          },
          "sectionTitle": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "sectionSubtitle": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "itemLimit": {
            "type": "number",
            "description": "AUTOMATIC mode only; ignored in MANUAL."
          },
          "ctaText": {
            "$ref": "#/components/schemas/LocalizedTextDto"
          },
          "ctaUrl": {
            "type": "string"
          },
          "visibleFrom": {
            "type": "string"
          },
          "visibleUntil": {
            "type": "string"
          },
          "displayOrder": {
            "type": "number"
          },
          "enabled": {
            "type": "boolean"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "Everyone",
              "AuthenticatedOnly",
              "AdminPreviewOnly"
            ]
          },
          "selectionMode": {
            "type": "string",
            "enum": [
              "MANUAL",
              "AUTOMATIC"
            ]
          },
          "items": {
            "description": "Manually-selected item ids (MANUAL mode). Bare ObjectIds — the target collection is inferred from sectionType; the board defines no per-entry type discriminator.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "filters": {
            "type": "object",
            "description": "Free-form AUTOMATIC-mode query params."
          },
          "configuration": {
            "type": "object",
            "description": "Free-form section-specific settings."
          }
        },
        "required": [
          "pageId",
          "sectionType",
          "displayOrder",
          "visibility",
          "selectionMode"
        ]
      },
      "PageSectionPublicResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sectionType": {
            "type": "string",
            "enum": [
              "HERO",
              "FEDERATION_STATS",
              "FEATURED_ATHLETES",
              "RESULTS_RANKINGS",
              "CLUBS",
              "UPCOMING_HIGHLIGHTS",
              "LATEST_NEWS",
              "PHOTO_GALLERY",
              "VIDEO_LIBRARY",
              "EXTERNAL_MEDIA",
              "SPONSORS",
              "PARTNERS",
              "NEWSLETTER_CTA"
            ]
          },
          "sectionTitle": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalizedTextDto"
              }
            ]
          },
          "sectionSubtitle": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalizedTextDto"
              }
            ]
          },
          "itemLimit": {
            "type": "object",
            "nullable": true
          },
          "ctaText": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LocalizedTextDto"
              }
            ]
          },
          "ctaUrl": {
            "type": "object",
            "nullable": true
          },
          "displayOrder": {
            "type": "number"
          },
          "selectionMode": {
            "type": "string",
            "enum": [
              "MANUAL",
              "AUTOMATIC"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "configuration": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "sectionType",
          "displayOrder",
          "selectionMode",
          "items"
        ]
      },
...
    }
  }
}