# api/src/modules/people-organizations/athletes/athletes.controller.ts
import { Body, Controller, Delete, Get, Header, Param, Post, Query } from '@nestjs/common';
import { ApiTags } from '@nestjs/swagger';
import { Types } from 'mongoose';
import { RequirePermission } from '../../../common/decorators/permissions.decorator.js';
import { CurrentUser } from '../../../common/decorators/current-user.decorator.js';
import { Public } from '../../../common/decorators/public.decorator.js';
import { PaginationQueryDto } from '../../../common/dto/pagination-query.dto.js';
import type { AuthenticatedUser } from '../../../common/interfaces/jwt-payload.interface.js';
import { AthletesService } from './athletes.service.js';
import { CreateAthleteDto } from './dto/create-athlete.dto.js';

/** Implements: athletes collection, Domain 2 — People & Organizations. */
@ApiTags('athletes')
@Controller('athletes')
export class AthletesController {
  constructor(private readonly service: AthletesService) {}

  @Post()
  @RequirePermission('athletes', 'Create')
  create(@Body() dto: CreateAthleteDto) {
    return this.service.create(dto);
  }

  @Get()
  @RequirePermission('athletes', 'Read')
  findAll() {
    return this.service.findAll();
  }

  /** Public athlete listing — public-safe shape only, structurally without
   *  `dateOfBirth` (`[SENSITIVE-MINOR]`, ADR-0028). Paginated via
   *  `page`/`limit` query params (2026-09-04, public-routes closure). */
  @Get('public')
  @Public()
  findAllPublic(@Query() query: PaginationQueryDto) {
    return this.service.findAllPublic(query.page, query.limit);
  }

  /** Declared before `:id` on purpose — Nest matches in declaration order,
   *  so a later `export` route would be swallowed by the id parameter and
   *  answer 404 for a path that exists. */
  @Get('export')
  @RequirePermission('athletes', 'Export')
  @Header('content-type', 'text/csv; charset=utf-8')
  @Header('content-disposition', 'attachment; filename="athletes.csv"')
  exportCsv() {
    return this.service.exportCsv();
  }

  @Get(':id')
  @RequirePermission('athletes', 'Read')
  findOne(@Param('id') id: string) {
    return this.service.findById(id);
  }

  @Delete(':id')
  @RequirePermission('athletes', 'Delete')
  remove(@Param('id') id: string, @CurrentUser() user: AuthenticatedUser) {
    return this.service.remove(id, new Types.ObjectId(user.userId));
  }
}